requirements.txt
# DMS - Document Management System
# Core dependencies
PySide6>=6.6.0
watchdog>=3.0.0
PyMuPDF>=1.23.0
requests>=2.31.0
Pillow>=10.0.0
PyYAML>=6.0.1
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-qt>=4.2.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
